move from repo to query dashboard requetes #5626
