+++ src/Repository/SignalementRepository.php
@@ -4,13 +4,10 @@ namespace App\Repository;
 
 use App\Dto\Api\Request\SignalementListQueryParams;
 use App\Dto\StatisticsFilters;
-use App\Entity\Affectation;
 use App\Entity\Commune;
-use App\Entity\EmailDeliveryIssue;
 use App\Entity\Enum\AffectationStatus;
 use App\Entity\Enum\CreationSource;
 use App\Entity\Enum\DesordreCritereZone;
-use App\Entity\Enum\PartnerType;
 use App\Entity\Enum\Qualification;
 use App\Entity\Enum\QualificationStatus;
 use App\Entity\Enum\SignalementStatus;
@@ -21,10 +18,7 @@ use App\Entity\Suivi;
 use App\Entity\Territory;
 use App\Entity\User;
 use App\Entity\UserSignalementSubscription;
-use App\Service\DashboardTabPanel\Kpi\CountAfermer;
 use App\Service\DashboardTabPanel\Kpi\CountNouveauxDossiers;
-use App\Service\DashboardTabPanel\TabDossier;
-use App\Service\DashboardTabPanel\TabQueryParameters;
 use App\Service\InjonctionBailleur\InjonctionBailleurService;
 use App\Service\Interconnection\Idoss\IdossService;
 use App\Service\ListFilters\SearchArchivedSignalement;
@@ -37,7 +31,6 @@ use App\Utils\CommuneHelper;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
-use Doctrine\DBAL\ArrayParameterType;
 use Doctrine\DBAL\Exception;
 use Doctrine\DBAL\LockMode;
 use Doctrine\ORM\NonUniqueResultException;
@@ -48,7 +41,6 @@ use Doctrine\ORM\QueryBuilder;
 use Doctrine\ORM\Tools\Pagination\Paginator;
 use Doctrine\ORM\TransactionRequiredException;
 use Doctrine\Persistence\ManagerRegistry;
-use Symfony\Bundle\SecurityBundle\Security;
 
 /**
  * @extends ServiceEntityRepository<Signalement>
@@ -65,7 +57,6 @@ class SignalementRepository extends ServiceEntityRepository
 
     public function __construct(
         ManagerRegistry $registry,
-        private readonly Security $security,
         private readonly PartnerAuthorizedResolver $partnerAuthorizedResolver,
     ) {
         parent::__construct($registry, Signalement::class);
@@ -1425,231 +1416,6 @@ class SignalementRepository extends ServiceEntityRepository
         return $qb->getQuery()->getResult();
     }
 
-    private function createSignalementQueryBuilder(
-        User $user,
-        ?SignalementStatus $signalementStatus = null,
-        ?AffectationStatus $affectationStatus = null,
-        ?bool $onlyWithoutSubscription = false,
-        ?TabQueryParameters $tabQueryParameters = null,
-    ): QueryBuilder {
-        $qb = $this->createQueryBuilder('s');
-
-        if (null !== $signalementStatus) {
-            $qb
-                ->andWhere('s.statut = :statut')
-                ->setParameter('statut', $signalementStatus);
-        }
-
-        if ($tabQueryParameters?->territoireId) {
-            $qb
-                ->andWhere('s.territory = :territoireId')
-                ->setParameter('territoireId', $tabQueryParameters->territoireId);
-        } elseif (!$user->isSuperAdmin()) {
-            $qb->andWhere('s.territory IN (:territories)')->setParameter('territories', $user->getPartnersTerritories());
-        }
-
-        if ($tabQueryParameters->createdFrom) {
-            if (CreationSource::CREATED_FROM_FORMULAIRE_USAGER === $tabQueryParameters->createdFrom) {
-                $qb->andWhere('s.creationSource IN (:creationSources)')
-                    ->setParameter('creationSources', CreationSource::getFormUsagerValues());
-            } elseif (CreationSource::CREATED_FROM_FORMULAIRE_PRO === $tabQueryParameters->createdFrom) {
-                $qb->andWhere('s.creationSource IN (:creationSources)')
-                    ->setParameter('creationSources', CreationSource::getFormProValues());
-            }
-        }
-
-        if (!empty($tabQueryParameters->partenairesId)) {
-            if (\in_array('AUCUN', $tabQueryParameters->partenairesId)) {
-                $qb->leftJoin('s.affectations', 'a')->andWhere('a.partner IS NULL');
-            } else {
-                $qb
-                    ->leftJoin('s.affectations', 'a')
-                    ->andWhere('a.partner IN (:partenairesId)')
-                    ->setParameter('partenairesId', $tabQueryParameters->partenairesId);
-            }
-        }
-
-        if ($affectationStatus) {
-            $qb->andWhere('a.statut = :affectationStatus');
-            $qb->setParameter('affectationStatus', $affectationStatus);
-        }
-
-        if ($onlyWithoutSubscription) {
-            $subquery = 'SELECT u FROM '.User::class.' u JOIN u.userPartners up JOIN up.partner p WHERE p IN (:partners)';
-            $qb
-                ->leftJoin('s.userSignalementSubscriptions', 'uss', 'WITH', 'uss.user IN ('.$subquery.')')
-                ->andWhere('uss.id IS NULL')
-                ->setParameter('partners', $user->getPartners());
-        }
-
-        return $qb;
-    }
-
-    /**
-     * @return TabDossier[]
-     */
-    public function findNewDossiersFrom(
-        ?SignalementStatus $signalementStatus = null,
-        ?AffectationStatus $affectationStatus = null,
-        ?TabQueryParameters $tabQueryParameters = null,
-    ): array {
-        /** @var User $user */
-        $user = $this->security->getUser();
-
-        $qb = $this->createSignalementQueryBuilder(
-            user: $user,
-            signalementStatus: $signalementStatus,
-            affectationStatus: $affectationStatus,
-            tabQueryParameters: $tabQueryParameters
-        );
-
-        if (CreationSource::CREATED_FROM_FORMULAIRE_PRO === $tabQueryParameters->createdFrom) {
-            $qb
-                ->leftJoin('s.createdBy', 'u')
-                ->leftJoin('u.userPartners', 'up')
-                ->leftJoin('up.partner', 'p');
-        }
-
-        $qb->select(
-            \sprintf(
-                'NEW %s(
-                    s.uuid,
-                    s.profileDeclarant,
-                    s.nomOccupant,
-                    s.prenomOccupant,
-                    s.reference,
-                    CONCAT_WS(\', \', s.adresseOccupant, CONCAT(s.cpOccupant, \' \', s.villeOccupant)),
-                    s.createdAt,'.
-                    (CreationSource::CREATED_FROM_FORMULAIRE_PRO === $tabQueryParameters->createdFrom
-                        ? 'CONCAT(UPPER(u.nom), \' \', u.prenom), p.nom,'
-                        : '\'\' , \'\' ,'
-                    ).
-                    'CASE
-                        WHEN s.isLogementSocial = true THEN \'PUBLIC\'
-                        ELSE \'PRIVÉ\'
-                    END,
-                    s.validatedAt
-                )',
-                TabDossier::class
-            )
-        );
-
-        if (null !== $tabQueryParameters
-            && in_array($tabQueryParameters->sortBy, ['createdAt', 'nomOccupant'], true)
-            && in_array($tabQueryParameters->orderBy, ['ASC', 'DESC', 'asc', 'desc'], true)
-        ) {
-            $qb->orderBy('s.'.$tabQueryParameters->sortBy, $tabQueryParameters->orderBy);
-        } else {
-            $qb->orderBy('s.createdAt', 'DESC');
-        }
-
-        $qb->setMaxResults(TabDossier::MAX_ITEMS_LIST);
-
-        return $qb->getQuery()->getResult();
-    }
-
-    /**
-     * @throws NonUniqueResultException
-     * @throws NoResultException
-     */
-    public function countNewDossiersFrom(
-        ?SignalementStatus $signalementStatus = null,
-        ?AffectationStatus $affectationStatus = null,
-        ?TabQueryParameters $tabQueryParameters = null,
-    ): int {
-        /** @var User $user */
-        $user = $this->security->getUser();
-
-        $qb = $this->createSignalementQueryBuilder(
-            user: $user,
-            signalementStatus: $signalementStatus,
-            affectationStatus: $affectationStatus,
-            tabQueryParameters: $tabQueryParameters
-        );
-
-        $qb->select('COUNT(s.id)');
-
-        return (int) $qb->getQuery()->getSingleScalarResult();
-    }
-
-    /**
-     * @return TabDossier[]
-     */
-    public function findDossiersNoAgentFrom(
-        ?AffectationStatus $affectationStatus = null,
-        ?TabQueryParameters $tabQueryParameters = null,
-        bool $isLimitApplied = true,
-    ): array {
-        /** @var User $user */
-        $user = $this->security->getUser();
-
-        $qb = $this->createSignalementQueryBuilder(
-            user: $user,
-            affectationStatus: $affectationStatus,
-            onlyWithoutSubscription: true,
-            tabQueryParameters: $tabQueryParameters
-        );
-
-        $qb->select(
-            \sprintf(
-                'NEW %s(
-                    s.uuid,
-                    s.profileDeclarant,
-                    s.nomOccupant,
-                    s.prenomOccupant,
-                    s.reference,
-                    CONCAT_WS(\', \', s.adresseOccupant, CONCAT(s.cpOccupant, \' \', s.villeOccupant)),
-                    s.createdAt,'.
-                    '\'\' , \'\' ,
-                    CASE
-                        WHEN s.isLogementSocial = true THEN \'PUBLIC\'
-                        ELSE \'PRIVÉ\'
-                    END,
-                    s.validatedAt
-                )',
-                TabDossier::class
-            )
-        );
-
-        if (null !== $tabQueryParameters
-            && in_array($tabQueryParameters->sortBy, ['createdAt', 'nomOccupant'], true)
-            && in_array($tabQueryParameters->orderBy, ['ASC', 'DESC', 'asc', 'desc'], true)
-        ) {
-            $qb->orderBy('s.'.$tabQueryParameters->sortBy, $tabQueryParameters->orderBy);
-        } else {
-            $qb->orderBy('s.createdAt', 'DESC');
-        }
-
-        if ($isLimitApplied) {
-            $qb->setMaxResults(TabDossier::MAX_ITEMS_LIST);
-        }
-
-        return $qb->getQuery()->getResult();
-    }
-
-    /**
-     * @throws NonUniqueResultException
-     * @throws NoResultException
-     */
-    public function countDossiersNoAgentFrom(
-        ?AffectationStatus $affectationStatus = null,
-        ?TabQueryParameters $tabQueryParameters = null,
-    ): int {
-        /** @var User $user */
-        $user = $this->security->getUser();
-
-        $qb = $this->createSignalementQueryBuilder(
-            user: $user,
-            affectationStatus: $affectationStatus,
-            onlyWithoutSubscription: true,
-            tabQueryParameters: $tabQueryParameters
-        );
-
-        $qb->select('COUNT(s.id)');
-
-        return (int) $qb->getQuery()->getSingleScalarResult();
-    }
-
     /**
      * @param array<int, mixed> $territories
      *
@@ -1707,636 +1473,6 @@ class SignalementRepository extends ServiceEntityRepository
         return $qb->getQuery()->getSingleResult();
     }
 
-    /**
-     * @return TabDossier[]
-     *
-     * @throws \DateMalformedStringException
-     */
-    public function findDossiersFermePartenaireTous(?TabQueryParameters $tabQueryParameters): array
-    {
-        /** @var User $user */
-        $user = $this->security->getUser();
-
-        $qb = $this->createSignalementQueryBuilder(
-            user: $user,
-            signalementStatus: SignalementStatus::ACTIVE,
-            tabQueryParameters: $tabQueryParameters
-        );
-
-        $qb->select("
-            s.uuid,
-            s.nomOccupant,
-            s.prenomOccupant,
-            s.reference,
-            CONCAT_WS(', ', s.adresseOccupant, CONCAT(s.cpOccupant, ' ', s.villeOccupant)) AS fullAddress,
-            MAX(a.answeredAt) AS lastClosedAt
-        ")
-            ->innerJoin('s.affectations', 'a')
-            ->groupBy('s.uuid, s.nomOccupant, s.prenomOccupant, s.reference, s.adresseOccupant, s.cpOccupant, s.villeOccupant')
-            ->having('COUNT(a.id) = SUM(CASE WHEN a.statut = :closed THEN 1 ELSE 0 END)')
-            ->setParameter('closed', AffectationStatus::CLOSED);
-
-        if (null !== $tabQueryParameters
-            && 'closedAt' === $tabQueryParameters->sortBy
-            && in_array($tabQueryParameters->orderBy, ['ASC', 'DESC', 'asc', 'desc'], true)
-        ) {
-            $qb->orderBy('MAX(a.answeredAt)', $tabQueryParameters->orderBy);
-        } else {
-            $qb->orderBy('MAX(a.answeredAt)', 'ASC');
-        }
-
-        $qb->setMaxResults(TabDossier::MAX_ITEMS_LIST);
-
-        $rows = $qb->getQuery()->getArrayResult();
-
-        return array_map(
-            static fn (array $row) => new TabDossier(
-                uuid: $row['uuid'],
-                nomOccupant: $row['nomOccupant'],
-                prenomOccupant: $row['prenomOccupant'],
-                reference: $row['reference'],
-                adresse: $row['fullAddress'],
-                clotureAt: $row['lastClosedAt'] ? new \DateTimeImmutable($row['lastClosedAt']) : null,
-            ),
-            $rows
-        );
-    }
-
-    public function countDossiersFermePartenaireTous(?TabQueryParameters $tabQueryParameters): int
-    {
-        /** @var User $user */
-        $user = $this->security->getUser();
-
-        $qb = $this->createSignalementQueryBuilder(
-            user: $user,
-            signalementStatus: SignalementStatus::ACTIVE,
-            tabQueryParameters: $tabQueryParameters
-        );
-
-        $em = $this->getEntityManager();
-
-        $existsAtLeastOneAffectation = $em->createQueryBuilder()
-            ->select('1')
-            ->from(Affectation::class, 'a1')
-            ->where('a1.signalement = s')
-            ->getDQL();
-
-        $existsAffectationNotClosed = $em->createQueryBuilder()
-            ->select('1')
-            ->from(Affectation::class, 'a2')
-            ->where('a2.signalement = s')
-            ->andWhere('a2.statut != :closed')
-            ->getDQL();
-
-        $qb->andWhere($qb->expr()->exists($existsAtLeastOneAffectation));
-        $qb->andWhere($qb->expr()->not($qb->expr()->exists($existsAffectationNotClosed)));
-
-        $qb->select('COUNT(DISTINCT s.id)');
-        $qb->setParameter('closed', AffectationStatus::CLOSED);
-
-        return (int) $qb->getQuery()->getSingleScalarResult();
-    }
-
-    /**
-     * @return TabDossier[]
-     *
-     * @throws \DateMalformedStringException
-     */
-    public function findDossiersFermePartenaireCommune(?TabQueryParameters $tabQueryParameters): array
-    {
-        /** @var User $user */
-        $user = $this->security->getUser();
-
-        $qb = $this->createSignalementQueryBuilder(
-            user: $user,
-            signalementStatus: SignalementStatus::ACTIVE,
-            tabQueryParameters: $tabQueryParameters
-        );
-
-        $qb->select("
-            s.uuid,
-            s.nomOccupant,
-            s.prenomOccupant,
-            s.reference,
-            CONCAT_WS(', ', s.adresseOccupant, CONCAT(s.cpOccupant, ' ', s.villeOccupant)) AS fullAddress,
-            MAX(a.answeredAt) AS lastClosedAt
-        ")
-            ->innerJoin('s.affectations', 'a')
-            ->innerJoin('a.partner', 'p')
-            ->andWhere('p.type = :partnerType')
-            ->andWhere('a.statut = :closed')
-            ->groupBy('s.uuid, s.nomOccupant, s.prenomOccupant, s.reference, s.adresseOccupant, s.cpOccupant, s.villeOccupant')
-            ->setParameter('closed', AffectationStatus::CLOSED)
-            ->setParameter('partnerType', PartnerType::COMMUNE_SCHS);
-
-        if (null !== $tabQueryParameters
-            && 'closedAt' === $tabQueryParameters->sortBy
-            && in_array($tabQueryParameters->orderBy, ['ASC', 'DESC', 'asc', 'desc'], true)
-        ) {
-            $qb->orderBy('MAX(a.answeredAt)', $tabQueryParameters->orderBy);
-        } else {
-            $qb->orderBy('MAX(a.answeredAt)', 'ASC');
-        }
-
-        $qb->setMaxResults(TabDossier::MAX_ITEMS_LIST);
-
-        $rows = $qb->getQuery()->getArrayResult();
-
-        return array_map(
-            static fn (array $row) => new TabDossier(
-                uuid: $row['uuid'],
-                nomOccupant: $row['nomOccupant'],
-                prenomOccupant: $row['prenomOccupant'],
-                reference: $row['reference'],
-                adresse: $row['fullAddress'],
-                clotureAt: $row['lastClosedAt'] ? new \DateTimeImmutable($row['lastClosedAt']) : null,
-            ),
-            $rows
-        );
-    }
-
-    public function countDossiersFermePartenaireCommune(?TabQueryParameters $tabQueryParameters): int
-    {
-        /** @var User $user */
-        $user = $this->security->getUser();
-
-        $qb = $this->createSignalementQueryBuilder(
-            user: $user,
-            signalementStatus: SignalementStatus::ACTIVE,
-            tabQueryParameters: $tabQueryParameters
-        );
-
-        $qb->select('COUNT(DISTINCT s.id)')
-            ->innerJoin('s.affectations', 'a')
-            ->innerJoin('a.partner', 'p')
-            ->andWhere('p.type = :partnerType')
-            ->andWhere('a.statut = :closed')
-            ->setParameter('partnerType', PartnerType::COMMUNE_SCHS)
-            ->setParameter('closed', AffectationStatus::CLOSED);
-
-        return (int) $qb->getQuery()->getSingleScalarResult();
-    }
-
-    /**
-     * @return TabDossier[]
-     *
-     * @throws \DateMalformedStringException
-     */
-    public function findDossiersDemandesFermetureByUsager(?TabQueryParameters $tabQueryParameters): array
-    {
-        /** @var User $user */
-        $user = $this->security->getUser();
-
-        $qb = $this->createSignalementQueryBuilder(
-            user: $user,
-            signalementStatus: SignalementStatus::ACTIVE,
-            tabQueryParameters: $tabQueryParameters
-        );
-
-        $qb->select("
-            s.uuid,
-            s.nomOccupant,
-            s.prenomOccupant,
-            s.reference,
-            CONCAT_WS(', ', s.adresseOccupant, CONCAT(s.cpOccupant, ' ', s.villeOccupant)) AS fullAddress,
-            MAX(su.createdAt) AS demandeFermetureUsagerAt,
-            DATEDIFF(CURRENT_DATE(), MAX(su.createdAt)) AS demandeFermetureUsagerDaysAgo,
-            CASE WHEN s.isNotOccupant = 1 THEN 'TIERS DÉCLARANT' ELSE 'OCCUPANT' END AS demandeFermetureUsagerProfileDeclarant
-        ")
-            ->innerJoin('s.suivis', 'su', 'WITH', 'su.category = :suivi_category_abandon_procedure')
-            ->andWhere('s.isUsagerAbandonProcedure = 1')
-            ->groupBy('s.uuid, s.nomOccupant, s.prenomOccupant, s.reference, s.adresseOccupant, s.cpOccupant, s.villeOccupant, s.isNotOccupant')
-            ->orderBy('MAX(su.createdAt)', 'DESC')
-            ->setParameter('suivi_category_abandon_procedure', SuiviCategory::DEMANDE_ABANDON_PROCEDURE);
-
-        if (null !== $tabQueryParameters
-            && 'demandeFermetureUsagerAt' === $tabQueryParameters->sortBy
-            && in_array($tabQueryParameters->orderBy, ['ASC', 'DESC', 'asc', 'desc'], true)
-        ) {
-            $qb->orderBy('MAX(su.createdAt)', $tabQueryParameters->orderBy);
-        } else {
-            $qb->orderBy('MAX(su.createdAt)', 'ASC');
-        }
-
-        $qb->setMaxResults($tabQueryParameters?->limit ?? TabDossier::MAX_ITEMS_LIST);
-
-        $rows = $qb->getQuery()->getArrayResult();
-
-        return array_map(
-            static fn (array $row) => new TabDossier(
-                uuid: $row['uuid'],
-                nomOccupant: $row['nomOccupant'] ?? null,
-                prenomOccupant: $row['prenomOccupant'] ?? null,
-                reference: $row['reference'] ?? null,
-                adresse: $row['fullAddress'] ?? null,
-                demandeFermetureUsagerDaysAgo: isset($row['demandeFermetureUsagerDaysAgo']) ? (int) $row['demandeFermetureUsagerDaysAgo'] : null,
-                demandeFermetureUsagerProfileDeclarant: $row['demandeFermetureUsagerProfileDeclarant'] ?? null,
-                demandeFermetureUsagerAt: null !== $row['demandeFermetureUsagerAt'] ? new \DateTimeImmutable($row['demandeFermetureUsagerAt']) : null
-            ),
-            $rows
-        );
-    }
-
-    /**
-     * @throws NonUniqueResultException
-     * @throws NoResultException
-     */
-    public function countDossiersDemandesFermetureByUsager(?TabQueryParameters $tabQueryParameters): int
-    {
-        /** @var User $user */
-        $user = $this->security->getUser();
-
-        $qb = $this->createSignalementQueryBuilder(
-            user: $user,
-            signalementStatus: SignalementStatus::ACTIVE,
-            tabQueryParameters: $tabQueryParameters
-        );
-
-        $qb
-            ->select('COUNT(DISTINCT s.uuid)')
-            ->innerJoin('s.suivis', 'su', 'WITH', 'su.category = :suivi_category_abandon_procedure')
-            ->andWhere('s.isUsagerAbandonProcedure = 1')
-            ->setParameter('suivi_category_abandon_procedure', SuiviCategory::DEMANDE_ABANDON_PROCEDURE);
-
-        return (int) $qb->getQuery()->getSingleScalarResult();
-    }
-
-    private function getBaseSignalementsAvecRelancesSansReponseSql(?array $territoriesIds): string
-    {
-        $clauseTerritoriesSi2 = '';
-        $clauseTerritoriesSi3 = '';
-        $clauseTerritoriesSi = '';
-        if (null !== $territoriesIds) {
-            $clauseTerritoriesSi2 = ' AND si2.territory_id IN (:territories_ids) ';
-            $clauseTerritoriesSi3 = ' AND si3.territory_id IN (:territories_ids) ';
-            $clauseTerritoriesSi = ' AND si.territory_id IN (:territories_ids) ';
-        }
-
-        return <<<SQL
-            FROM (
-                SELECT
-                    s.signalement_id,
-                    MIN(s.created_at) AS first_relance_at,
-                    COUNT(*) AS nb_relances
-                FROM suivi s
-                WHERE s.category = 'ASK_FEEDBACK_SENT'
-                  AND EXISTS (
-                    SELECT 1 FROM signalement si2
-                    WHERE si2.id = s.signalement_id
-                      AND si2.statut = 'ACTIVE'
-                      $clauseTerritoriesSi2
-                  )
-                GROUP BY s.signalement_id
-                HAVING COUNT(*) >= 3
-            ) relances_usager
-            INNER JOIN signalement si ON si.id = relances_usager.signalement_id
-            INNER JOIN (
-                SELECT
-                    s.signalement_id,
-                    MAX(s.created_at) AS shared_usager_at,
-                    MAX(s.type) AS type
-                FROM suivi s
-                WHERE s.is_public = 1
-                  AND EXISTS (
-                    SELECT 1 FROM signalement si3
-                    WHERE si3.id = s.signalement_id
-                      AND si3.statut = 'ACTIVE'
-                      $clauseTerritoriesSi3
-                  )
-                GROUP BY s.signalement_id
-            ) last_usager_suivi ON last_usager_suivi.signalement_id = si.id
-            WHERE
-                si.statut = 'ACTIVE'
-                AND NOT EXISTS (
-                    SELECT 1
-                    FROM suivi s2
-                    WHERE s2.signalement_id = relances_usager.signalement_id
-                      AND s2.type = 2
-                      AND s2.created_at > relances_usager.first_relance_at
-                )
-                $clauseTerritoriesSi
-        SQL;
-    }
-
-    /**
-     * @return TabDossier[]
-     *
-     * @throws \DateMalformedStringException
-     * @throws Exception
-     */
-    public function findSignalementsAvecRelancesSansReponse(TabQueryParameters $tabQueryParameters): array
-    {
-        $conn = $this->getEntityManager()->getConnection();
-
-        $sql = <<<SQL
-            SELECT
-                si.uuid,
-                si.id,
-                si.reference,
-                si.nom_occupant,
-                si.prenom_occupant,
-                CONCAT_WS(', ', si.adresse_occupant, CONCAT(si.cp_occupant, ' ', si.ville_occupant)) AS fullAddress,
-                relances_usager.nb_relances,
-                relances_usager.first_relance_at,
-                last_usager_suivi.shared_usager_at AS last_suivi_shared_usager_at,
-                last_usager_suivi.type AS last_suivi_type
-        SQL;
-        $territoriesIds = $tabQueryParameters->territoireId ? [$tabQueryParameters->territoireId] : null;
-        $sql .= $this->getBaseSignalementsAvecRelancesSansReponseSql($territoriesIds);
-
-        if ('ASC' === $tabQueryParameters->orderBy && 'nbRelanceFeedbackUsager' === $tabQueryParameters->sortBy) {
-            $sql .= ' ORDER BY relances_usager.nb_relances ASC, relances_usager.first_relance_at DESC LIMIT 5';
-        } else {
-            $sql .= ' ORDER BY relances_usager.nb_relances DESC, relances_usager.first_relance_at DESC LIMIT 5';
-        }
-
-        if (null === $territoriesIds) {
-            $rows = $conn->executeQuery($sql)->fetchAllAssociative();
-        } else {
-            $rows = $conn->executeQuery($sql,
-                ['territories_ids' => $territoriesIds],
-                ['territories_ids' => ArrayParameterType::INTEGER]
-            )->fetchAllAssociative();
-        }
-
-        return array_map(/**
-         * @throws \DateMalformedStringException
-         */ static function (array $row): TabDossier {
-            return new TabDossier(
-                uuid: $row['uuid'],
-                nomOccupant: $row['nom_occupant'],
-                prenomOccupant: $row['prenom_occupant'],
-                reference: $row['reference'],
-                adresse: $row['fullAddress'],
-                nbRelanceDossier: (int) $row['nb_relances'],
-                premiereRelanceDossierAt: new \DateTimeImmutable($row['first_relance_at']),
-                dernierSuiviPublicAt: new \DateTimeImmutable($row['last_suivi_shared_usager_at']),
-                dernierTypeSuivi: (string) $row['last_suivi_type'],
-            );
-        }, $rows);
-    }
-
-    public function countSignalementsAvecRelancesSansReponse(TabQueryParameters $tabQueryParameters): int
-    {
-        $conn = $this->getEntityManager()->getConnection();
-        $territoriesIds = $tabQueryParameters->territoireId ? [$tabQueryParameters->territoireId] : null;
-        $sql = 'SELECT COUNT(*) FROM (SELECT relances_usager.signalement_id '.$this->getBaseSignalementsAvecRelancesSansReponseSql($territoriesIds).') AS signalements_count';
-
-        if (null === $territoriesIds) {
-            return (int) $conn->executeQuery($sql)->fetchOne();
-        }
-
-        return (int) $conn->executeQuery($sql,
-            ['territories_ids' => $territoriesIds],
-            ['territories_ids' => ArrayParameterType::INTEGER]
-        )->fetchOne();
-    }
-
-    /**
-     * @return int[]
-     */
-    public function getSignalementsIdAvecRelancesSansReponse(User $user): array
-    {
-        $conn = $this->getEntityManager()->getConnection();
-        $territoriesIds = [];
-        if (!$user->isSuperAdmin()) {
-            foreach ($user->getPartnersTerritories() as $territory) {
-                $territoriesIds[] = $territory->getId();
-            }
-        }
-        $territoriesIds = empty($territoriesIds) ? null : $territoriesIds;
-        $sql = 'SELECT si.id '.$this->getBaseSignalementsAvecRelancesSansReponseSql($territoriesIds);
-
-        if (null === $territoriesIds) {
-            return array_map('intval', $conn->executeQuery($sql)->fetchFirstColumn());
-        }
-
-        return array_map('intval', $conn->executeQuery($sql,
-            ['territories_ids' => $territoriesIds],
-            ['territories_ids' => ArrayParameterType::INTEGER]
-        )->fetchFirstColumn());
-    }
-
-    public function countAllDossiersAferme(User $user, ?TabQueryParameters $params): CountAfermer
-    {
-        return new CountAfermer(
-            countDemandesFermetureByUsager: $this->countDossiersDemandesFermetureByUsager($params),
-            countDossiersRelanceSansReponse: $this->countSignalementsAvecRelancesSansReponse($params),
-            countDossiersFermePartenaireTous: $this->countDossiersFermePartenaireTous($params),
-            countDossiersFermeCommune: $this->countDossiersFermePartenaireCommune($params),
-        );
-    }
-
-    /**
-     * @return array<int, array<string, array<int|string|null>|int|string>|string>
-     */
-    private function getBaseSignalementsSansSuiviPartenaireDepuis60JSql(
-        User $user,
-        TabQueryParameters $params,
-        ?bool $withJoins = false,
-    ): array {
-        /** @var SuiviRepository $suiviRepository */
-        $suiviRepository = $this->_em->getRepository(Suivi::class);
-        $excludedIds = array_merge(
-            $this->getSignalementsIdAvecRelancesSansReponse($user),
-            $suiviRepository->getSignalementsIdWithSuivisUsagerOrPoursuiteWithAskFeedbackBefore($user, $params)
-        );
-        $categories = [
-            'MESSAGE_PARTNER',
-            'SIGNALEMENT_EDITED_BO',
-            'SIGNALEMENT_IS_ACTIVE',
-            'SIGNALEMENT_IS_REOPENED',
-            'INTERVENTION_IS_CREATED',
-            'INTERVENTION_IS_CANCELED',
-            'INTERVENTION_IS_ABORTED',
-            'INTERVENTION_HAS_CONCLUSION',
-            'INTERVENTION_HAS_CONCLUSION_EDITED',
-            'INTERVENTION_IS_RESCHEDULED',
-            'INTERVENTION_IS_DONE',
-            'INTERVENTION_CONTROLE_IS_CREATED',
-            'INTERVENTION_CONTROLE_IS_RESCHEDULED',
-            'INTERVENTION_CONTROLE_IS_DONE',
-            'INTERVENTION_ARRETE_IS_CREATED',
-            'INTERVENTION_ARRETE_IS_RESCHEDULED',
-            'NEW_DOCUMENT',
-            'AFFECTATION_IS_CLOSED',
-        ];
-
-        $paramsToBind = [];
-        $types = [];
-        $categoryList = "'".implode("','", $categories)."'";
-        $sql = <<<SQL
-            FROM signalement si
-            INNER JOIN suivi s ON s.signalement_id = si.id
-        SQL;
-        if ($withJoins) {
-            $sql .= <<<SQL
-                LEFT JOIN user u ON u.id = s.created_by_id
-                LEFT JOIN user_partner up ON up.user_id = u.id
-                LEFT JOIN partner p ON p.id = up.partner_id
-            SQL;
-        }
-        if ($user->isPartnerAdmin() || $user->isUserPartner() || ($params->partners && count($params->partners) > 0)) {
-            $sql .= <<<SQL
-                LEFT JOIN affectation aff ON aff.signalement_id = si.id
-            SQL;
-        }
-
-        $sql .= <<<SQL
-            WHERE s.category IN ($categoryList)
-            AND s.created_at = (
-                SELECT MAX(s2.created_at)
-                FROM suivi s2
-                WHERE s2.signalement_id = si.id
-                AND s2.category IN ($categoryList)
-            )
-            AND s.created_at < :dateLimit
-            AND si.statut = 'ACTIVE'
-        SQL;
-
-        if ($user->isPartnerAdmin() || $user->isUserPartner()) {
-            $sql .= ' AND aff.partner_id IN (:partners)';
-            $sql .= ' AND aff.statut IN (\'EN_COURS\', \'NOUVEAU\')';
-            $paramsToBind['partners'] = array_map(
-                static fn ($partner) => $partner->getId(),
-                $user->getPartners()->toArray()
-            );
-            $types['partners'] = ArrayParameterType::INTEGER;
-        }
-
-        if ($params->territoireId) {
-            $sql .= ' AND si.territory_id = '.$params->territoireId;
-        } elseif (!$user->isSuperAdmin()) {
-            $sql .= ' AND si.territory_id IN ('.implode(',', array_keys($user->getPartnersTerritories())).')';
-        }
-
-        if ($params->mesDossiersAverifier && '1' === $params->mesDossiersAverifier) {
-            $sql .= ' AND EXISTS (
-            SELECT 1
-            FROM user_signalement_subscription uss
-            WHERE uss.signalement_id = si.id
-              AND uss.user_id = '.$user->getId().'
-        )';
-        }
-
-        $paramsToBind['dateLimit'] = (new \DateTimeImmutable('-60 days'))->format('Y-m-d H:i:s');
-
-        if ($params->partners && count($params->partners) > 0) {
-            $sql .= ' AND aff.partner_id IN (:partnersId)';
-            $paramsToBind['partnersId'] = $params->partners;
-            $types['partnersId'] = ArrayParameterType::INTEGER;
-        }
-
-        if ($params->queryCommune) {
-            $listCity = [$params->queryCommune];
-            if (isset(CommuneHelper::COMMUNES_ARRONDISSEMENTS[$params->queryCommune])) {
-                $listCity = array_merge($listCity, CommuneHelper::COMMUNES_ARRONDISSEMENTS[$params->queryCommune]);
-            }
-            $sql .= ' AND (si.cp_occupant IN (:cities) OR si.ville_occupant IN (:cities))';
-            $paramsToBind['cities'] = $listCity;
-            $types['cities'] = ArrayParameterType::STRING;
-        }
-
-        if (!empty($excludedIds)) {
-            $sql .= ' AND si.id NOT IN (:excludedIds)';
-            $paramsToBind['excludedIds'] = $excludedIds;
-            $types['excludedIds'] = ArrayParameterType::INTEGER;
-        }
-
-        return [$sql, $paramsToBind, $types];
-    }
-
-    public function countSignalementsSansSuiviPartenaireDepuis60Jours(User $user, ?TabQueryParameters $params): int
-    {
-        $conn = $this->getEntityManager()->getConnection();
-
-        $sql = 'SELECT COUNT(DISTINCT si.id) ';
-        /** @var string $sqlPrincipal */
-        /** @var array<mixed> $paramsToBind */
-        /** @var array<mixed> $types */
-        [$sqlPrincipal, $paramsToBind, $types] = $this->getBaseSignalementsSansSuiviPartenaireDepuis60JSql($user, $params);
-        $sql .= $sqlPrincipal;
-
-        return (int) $conn->executeQuery($sql, $paramsToBind, $types)->fetchOne();
-    }
-
-    /**
-     * @return int[]
-     */
-    public function getSignalementsIdSansSuiviPartenaireDepuis60Jours(User $user, TabQueryParameters $params): array
-    {
-        $conn = $this->getEntityManager()->getConnection();
-
-        $sql = 'SELECT DISTINCT si.id ';
-        /** @var string $sqlPrincipal */
-        /** @var array<mixed> $paramsToBind */
-        /** @var array<mixed> $types */
-        [$sqlPrincipal, $paramsToBind, $types] = $this->getBaseSignalementsSansSuiviPartenaireDepuis60JSql($user, $params);
-        $sql .= $sqlPrincipal;
-
-        return array_map('intval', $conn->executeQuery($sql, $paramsToBind, $types)->fetchFirstColumn());
-    }
-
-    /**
-     * @return array<int, array<string, mixed>>
-     */
-    public function findSignalementsSansSuiviPartenaireDepuis60Jours(User $user, ?TabQueryParameters $params): array
-    {
-        $conn = $this->getEntityManager()->getConnection();
-
-        $sql = <<<SQL
-        SELECT
-            si.id,
-            si.uuid AS uuid,
-            si.reference AS reference,
-            si.nom_occupant AS nomOccupant,
-            si.prenom_occupant AS prenomOccupant,
-            CONCAT_WS(', ', si.adresse_occupant, CONCAT(si.cp_occupant, ' ', si.ville_occupant)) AS adresse,
-            MAX(s.created_at) AS dernierSuiviAt,
-            DATEDIFF(CURRENT_DATE(), MAX(s.created_at)) AS nbJoursDepuisDernierSuivi,
-            MAX(s.category) AS suiviCategory,
-            MAX(p.nom) AS derniereActionPartenaireNom,
-            MAX(u.nom) AS derniereActionPartenaireNomAgent,
-            MAX(u.prenom) AS derniereActionPartenairePrenomAgent
-        SQL;
-        /** @var string $sqlPrincipal */
-        /** @var array<mixed> $paramsToBind */
-        /** @var array<mixed> $types */
-        [$sqlPrincipal, $paramsToBind, $types] = $this->getBaseSignalementsSansSuiviPartenaireDepuis60JSql($user, $params, true);
-        $sql .= $sqlPrincipal;
-        $sql .= ' GROUP BY si.id, si.uuid, si.reference, si.nom_occupant, si.prenom_occupant, si.adresse_occupant, si.cp_occupant, si.ville_occupant';
-
-        if ($params && in_array($params->sortBy, ['createdAt'], true)
-            && in_array($params->orderBy, ['ASC', 'DESC', 'asc', 'desc'], true)
-        ) {
-            $sql .= ' ORDER BY MAX(s.created_at) '.$params->orderBy;
-        } else {
-            $sql .= ' ORDER BY MAX(s.created_at) ASC';
-        }
-
-        $sql .= ' LIMIT '.TabDossier::MAX_ITEMS_LIST;
-
-        return $conn->executeQuery($sql, $paramsToBind, $types)->fetchAllAssociative();
-    }
-
-    /**
-     * @return string[]
-     */
-    public function getSignalementsUuidSansAgent(TabQueryParameters $params): array
-    {
-        /** @var User $user */
-        $user = $this->security->getUser();
-        if (null === $params->territoireId) {
-            $params->partenairesId = $user->getPartners()
-                ->map(static fn ($partner) => $partner->getId())
-                ->toArray();
-        }
-
-        $signalements = $this->findDossiersNoAgentFrom(AffectationStatus::ACCEPTED, $params, false);
-
-        return array_map(static fn (TabDossier $dossier) => $dossier->uuid, $signalements);
-    }
-
     public function trimFields(): void
     {
         $connection = $this->getEntityManager()->getConnection();
@@ -2353,139 +1489,6 @@ class SignalementRepository extends ServiceEntityRepository
             nom_proprio = TRIM(REPLACE(nom_proprio, UNHEX("C2A0"), " "))';
 
         $connection->prepare($sql)->executeStatement();
-    }
-
-    private function buildBaseQbForNonDeliverable(User $user, ?TabQueryParameters $params): QueryBuilder
-    {
-        $qb = $this->_em->createQueryBuilder();
-
-        $qb->from(Signalement::class, 's')
-            ->innerJoin(
-                EmailDeliveryIssue::class,
-                'edi',
-                'WITH',
-                $qb->expr()->orX(
-                    $qb->expr()->eq('s.mailOccupant', 'edi.email'),
-                    $qb->expr()->eq('s.mailDeclarant', 'edi.email')
-                )
-            )
-            ->leftJoin('s.affectations', 'aff')
-            ->where('s.statut IN (:statutList)')
-            ->setParameter('statutList', [SignalementStatus::NEED_VALIDATION, SignalementStatus::ACTIVE]);
-
-        if ($user->isPartnerAdmin() || $user->isUserPartner()) {
-            $existsAffectation = $this->_em->createQueryBuilder()
-                ->select('1')
-                ->from(Affectation::class, 'af')
-                ->where('af.signalement = s')
-                ->andWhere('af.partner IN (:partners)')
-                ->andWhere('af.statut = :affectationStatus')
-                ->getDQL();
-            $qb->andWhere($qb->expr()->exists($existsAffectation))
-                ->setParameter('partners', $user->getPartners())
-                ->setParameter('affectationStatus', AffectationStatus::ACCEPTED);
-        }
-
-        if ($params?->territoireId) {
-            $qb
-                ->andWhere('s.territory = :territoireId')
-                ->setParameter('territoireId', $params->territoireId);
-        } elseif (!$user->isSuperAdmin()) {
-            $qb
-                ->andWhere('s.territory IN (:territories)')
-                ->setParameter('territories', $user->getPartnersTerritories());
-        }
-
-        if ($params && $params->mesDossiersAverifier && '1' === $params->mesDossiersAverifier) {
-            $existsSubscription = $this->_em->createQueryBuilder()
-                ->select('1')
-                ->from(UserSignalementSubscription::class, 'uss')
-                ->where('uss.signalement = s')
-                ->andWhere('uss.user = :currentUser')
-                ->getDQL();
-            $qb->andWhere($qb->expr()->exists($existsSubscription))
-                ->setParameter('currentUser', $user);
-        }
-
-        if ($params && $params->queryCommune) {
-            $query = '%'.$params->queryCommune.'%';
-            $qb
-                ->andWhere(
-                    $qb->expr()->orX(
-                        $qb->expr()->like('s.cpOccupant', ':query'),
-                        $qb->expr()->like('s.villeOccupant', ':query')
-                    )
-                )
-                ->setParameter('query', '%'.$query.'%');
-        }
-
-        if ($params && $params->partners && count($params->partners) > 0) {
-            $qb->andWhere('aff.partner IN (:partnersId)')
-                ->setParameter('partnersId', $params->partners);
-        }
-
-        return $qb;
-    }
-
-    /**
-     * @return array<int, array<string, mixed>>
-     */
-    public function findActiveSignalementsWithInvalidEmails(User $user, ?TabQueryParameters $params): array
-    {
-        $qb = $this->buildBaseQbForNonDeliverable($user, $params);
-
-        $qb->select(
-            's.uuid AS uuid,
-            s.nomOccupant AS nomOccupant,
-            s.prenomOccupant AS prenomOccupant,
-            s.reference AS reference,
-            CONCAT_WS(\', \', s.adresseOccupant, CONCAT(s.cpOccupant, \' \', s.villeOccupant)) AS adresse,
-            s.createdAt AS createdAt,
-            s.lastSuiviAt AS dernierSuiviAt,
-            s.lastSuiviBy AS derniereActionPartenaireNom,
-            CASE
-                WHEN (FIND_IN_SET(s.mailOccupant, GROUP_CONCAT(DISTINCT edi.email)) > 0
-                    AND FIND_IN_SET(s.mailDeclarant, GROUP_CONCAT(DISTINCT edi.email)) > 0)
-                    THEN \'Occupant et Tiers\'
-                WHEN FIND_IN_SET(s.mailOccupant, GROUP_CONCAT(DISTINCT edi.email)) > 0
-                    THEN \'Occupant\'
-                WHEN FIND_IN_SET(s.mailDeclarant, GROUP_CONCAT(DISTINCT edi.email)) > 0
-                    THEN \'Tiers\'
-                ELSE \'\'
-            END AS profilNonDeliverable'
-        );
-
-        if ($params && in_array($params->sortBy, ['createdAt', 'nomOccupant'], true)
-            && in_array($params->orderBy, ['ASC', 'DESC', 'asc', 'desc'], true)
-        ) {
-            $qb->orderBy($params->sortBy, $params->orderBy);
-        } else {
-            $qb->orderBy('createdAt', 'DESC');
-        }
-
-        $qb->setMaxResults(TabDossier::MAX_ITEMS_LIST);
-        $qb->groupBy('s.id');
-
-        return $qb->getQuery()->getArrayResult();
-    }
-
-    /**
-     * @return array<int>
-     */
-    public function findIdsNonDeliverableSignalements(User $user, ?TabQueryParameters $params): array
-    {
-        $qb = $this->buildBaseQbForNonDeliverable($user, $params);
-        $qb->select('s.id')->groupBy('s.id');
-
-        return $qb->getQuery()->getSingleColumnResult();
-    }
-
-    public function countNonDeliverableSignalements(User $user, ?TabQueryParameters $params): int
-    {
-        $qb = $this->buildBaseQbForNonDeliverable($user, $params);
-        $qb->select('COUNT(DISTINCT s.id)');
-
-        return (int) $qb->getQuery()->getSingleScalarResult();
     }
 
     /**
@@ -2519,29 +1522,6 @@ class SignalementRepository extends ServiceEntityRepository
             ->setParameter('injonctionCategories', SuiviCategory::injonctionBailleurCategories())
             ->getQuery()
             ->getOneOrNullResult();
-    }
-
-    public function countInjonctions(
-        User $user,
-        ?TabQueryParameters $params,
-    ): int {
-        $qb = $this->createQueryBuilder('s')
-            ->where('s.statut = :statut')
-            ->setParameter('statut', SignalementStatus::INJONCTION_BAILLEUR);
-
-        $qb->select('COUNT(s.id)');
-
-        if ($params?->territoireId) {
-            $qb
-                ->andWhere('s.territory = :territoireId')
-                ->setParameter('territoireId', $params->territoireId);
-        } elseif (!$user->isSuperAdmin()) {
-            $qb
-                ->andWhere('s.territory IN (:territories)')
-                ->setParameter('territories', $user->getPartnersTerritories());
-        }
-
-        return (int) $qb->getQuery()->getSingleScalarResult();
     }
 
     /**
