+++ tests/Functional/Repository/SuiviRepositoryTest.php
@@ -2,14 +2,10 @@
 
 namespace App\Tests\Functional\Repository;
 
-use App\Entity\Enum\NotificationType;
-use App\Entity\Notification;
 use App\Entity\Signalement;
 use App\Entity\Suivi;
-use App\Entity\User;
 use App\Repository\SignalementRepository;
 use App\Repository\SuiviRepository;
-use App\Service\DashboardTabPanel\TabQueryParameters;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
@@ -76,215 +72,5 @@ class SuiviRepositoryTest extends KernelTestCase
         $signalementRepository = $this->entityManager->getRepository(Signalement::class);
         $signalement = $signalementRepository->findOneBy(['id' => $result[0]]);
         $this->assertEquals('2022-8', $signalement->getReference());
-    }
-
-    public function testFindLastSignalementsWithUserSuivi(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        $territory = null;
-        $limit = 5;
-        $result = $this->suiviRepository->findLastSignalementsWithUserSuivi($user, $territory, $limit);
-        $this->assertIsArray($result);
-        $this->assertLessThanOrEqual($limit, count($result));
-        if (count($result) > 0) {
-            $row = $result[0];
-            $this->assertArrayHasKey('reference', $row);
-            $this->assertArrayHasKey('nomOccupant', $row);
-            $this->assertArrayHasKey('prenomOccupant', $row);
-            $this->assertArrayHasKey('adresseOccupant', $row);
-            $this->assertArrayHasKey('uuid', $row);
-            $this->assertArrayHasKey('statut', $row);
-            $this->assertArrayHasKey('suiviCreatedAt', $row);
-            $this->assertArrayHasKey('suiviCategory', $row);
-            $this->assertArrayHasKey('suiviIsPublic', $row);
-            $this->assertArrayHasKey('hasNewerSuivi', $row);
-        }
-    }
-
-    public function testCountSuivisUsagersWithoutAskFeedbackBefore(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        $tabQueryParameter = new TabQueryParameters(
-            mesDossiersMessagesUsagers: '0',
-            sortBy: 'createdAt',
-            orderBy: 'DESC',
-        );
-
-        $result = $this->suiviRepository->countSuivisUsagersWithoutAskFeedbackBefore($user, $tabQueryParameter);
-        $this->assertIsInt($result);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testFindSuivisUsagersWithoutAskFeedbackBefore(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        $tabQueryParameter = new TabQueryParameters(
-            mesDossiersMessagesUsagers: '0',
-            sortBy: 'createdAt',
-            orderBy: 'DESC',
-        );
-        $result = $this->suiviRepository->findSuivisUsagersWithoutAskFeedbackBefore($user, $tabQueryParameter);
-        $this->assertIsArray($result);
-
-        if (count($result) > 0) {
-            $this->assertIsArray($result[0]);
-            $this->assertArrayHasKey('reference', $result[0]);
-            $this->assertEquals('2022-4', $result[0]['reference']);
-        }
-    }
-
-    public function testCountSuivisPostCloture(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        $tabQueryParameter = new TabQueryParameters(
-            mesDossiersMessagesUsagers: '0',
-            sortBy: 'createdAt',
-            orderBy: 'DESC',
-        );
-        $result = $this->suiviRepository->countSuivisPostCloture($user, $tabQueryParameter);
-        $this->assertIsInt($result);
-        $this->assertEquals(1, $result);
-    }
-
-    public function testFindSuivisPostCloture(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        $tabQueryParameter = new TabQueryParameters(
-            mesDossiersMessagesUsagers: '0',
-            sortBy: 'createdAt',
-            orderBy: 'DESC',
-        );
-        $result = $this->suiviRepository->findSuivisPostCloture($user, $tabQueryParameter);
-        $this->assertIsArray($result);
-
-        if (count($result) > 0) {
-            $this->assertIsArray($result[0]);
-            $this->assertArrayHasKey('reference', $result[0]);
-            $this->assertEquals('2022-2', $result[0]['reference']);
-        }
-    }
-
-    public function testCountSuivisPostClotureSeen(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        /** @var Signalement $signalementWithSuiviPostCloture */
-        $signalementWithSuiviPostCloture = $this->entityManager->getRepository(Signalement::class)->findOneBy(['reference' => '2022-2']);
-        /** @var Notification $notification */
-        $notification = $this->entityManager->getRepository(Notification::class)->findOneBy([
-            'signalement' => $signalementWithSuiviPostCloture,
-            'user' => $user,
-            'type' => NotificationType::NOUVEAU_SUIVI,
-        ]);
-        $notification->setSeenAt(new \DateTimeImmutable());
-        $this->entityManager->persist($notification);
-        $this->entityManager->flush();
-
-        $tabQueryParameter = new TabQueryParameters(
-            mesDossiersMessagesUsagers: '0',
-            sortBy: 'createdAt',
-            orderBy: 'DESC',
-        );
-        $result = $this->suiviRepository->countSuivisPostCloture($user, $tabQueryParameter);
-        $this->assertIsInt($result);
-        $this->assertEquals(0, $result);
-    }
-
-    public function testCountSuivisPostClotureDeleted(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        /** @var Signalement $signalementWithSuiviPostCloture */
-        $signalementWithSuiviPostCloture = $this->entityManager->getRepository(Signalement::class)->findOneBy(['reference' => '2022-2']);
-        /** @var Notification $notification */
-        $notification = $this->entityManager->getRepository(Notification::class)->findOneBy([
-            'signalement' => $signalementWithSuiviPostCloture,
-            'user' => $user,
-            'type' => NotificationType::NOUVEAU_SUIVI,
-        ]);
-        $notification->setDeleted(true);
-        $this->entityManager->persist($notification);
-        $this->entityManager->flush();
-
-        $tabQueryParameter = new TabQueryParameters(
-            mesDossiersMessagesUsagers: '0',
-            sortBy: 'createdAt',
-            orderBy: 'DESC',
-        );
-        $result = $this->suiviRepository->countSuivisPostCloture($user, $tabQueryParameter);
-        $this->assertIsInt($result);
-        $this->assertEquals(0, $result);
-    }
-
-    public function testCountSuivisPostClotureNoNotification(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        /** @var Signalement $signalementWithSuiviPostCloture */
-        $signalementWithSuiviPostCloture = $this->entityManager->getRepository(Signalement::class)->findOneBy(['reference' => '2022-2']);
-        /** @var Notification $notification */
-        $notification = $this->entityManager->getRepository(Notification::class)->findOneBy([
-            'signalement' => $signalementWithSuiviPostCloture,
-            'user' => $user,
-            'type' => NotificationType::NOUVEAU_SUIVI,
-        ]);
-        $this->entityManager->remove($notification);
-        $this->entityManager->flush();
-
-        $tabQueryParameter = new TabQueryParameters(
-            mesDossiersMessagesUsagers: '0',
-            sortBy: 'createdAt',
-            orderBy: 'DESC',
-        );
-        $result = $this->suiviRepository->countSuivisPostCloture($user, $tabQueryParameter);
-        $this->assertIsInt($result);
-        $this->assertEquals(0, $result);
-    }
-
-    public function testCountSuivisUsagerOrPoursuiteWithAskFeedbackBefore(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        $tabQueryParameter = new TabQueryParameters(
-            mesDossiersMessagesUsagers: '1',
-            sortBy: 'createdAt',
-            orderBy: 'DESC',
-        );
-        $result = $this->suiviRepository->countSuivisUsagerOrPoursuiteWithAskFeedbackBefore($user, $tabQueryParameter);
-        $this->assertIsInt($result);
-        $this->assertGreaterThanOrEqual(0, $result);
-    }
-
-    public function testFindSuivisUsagerOrPoursuiteWithAskFeedbackBefore(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        $tabQueryParameter = new TabQueryParameters(
-            mesDossiersMessagesUsagers: '1',
-            sortBy: 'createdAt',
-            orderBy: 'DESC',
-        );
-        $result = $this->suiviRepository->findSuivisUsagerOrPoursuiteWithAskFeedbackBefore($user, $tabQueryParameter);
-        $this->assertIsArray($result);
-
-        if (count($result) > 0) {
-            $this->assertInstanceOf(Suivi::class, $result[0]);
-        }
-    }
-
-    public function testCountAllMessagesUsagers(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        $params = new TabQueryParameters();
-        $result = $this->suiviRepository->countAllMessagesUsagers($user, $params);
-        $this->assertIsObject($result);
-        $this->assertTrue(method_exists($result, 'total'));
-    }
-
-    public function testFindLastSignalementsWithOtherUserSuivi(): void
-    {
-        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => self::USER_ADMIN]);
-        $tabQueryParameter = new TabQueryParameters(
-            mesDossiersActiviteRecente: '1',
-            sortBy: 'createdAt',
-            orderBy: 'DESC',
-        );
-        $result = $this->suiviRepository->findLastSignalementsWithOtherUserSuivi($user, $tabQueryParameter);
-        $this->assertIsArray($result);
     }
 }
