+++ tests/Unit/Service/DashboardTabPanel/TabDataManagerTest.php
@@ -10,10 +10,15 @@ use App\Entity\Enum\SignalementStatus;
 use App\Entity\Enum\SuiviCategory;
 use App\Entity\User;
 use App\Repository\JobEventRepository;
-use App\Repository\PartnerRepository;
+use App\Repository\Query\Dashboard\DossiersActiviteRecenteQuery;
+use App\Repository\Query\Dashboard\DossiersAvecRelanceSansReponseQuery;
+use App\Repository\Query\Dashboard\DossiersQuery;
+use App\Repository\Query\Dashboard\DossiersSansSuivisPartenaireQuery;
+use App\Repository\Query\Dashboard\DossiersSuivisUsagerQuery;
+use App\Repository\Query\Dashboard\DossiersUndeliverableEmailQuery;
+use App\Repository\Query\Dashboard\KpiQuery;
 use App\Repository\Query\Dashboard\SignalementsSansAffectationAccepteeQuery;
 use App\Repository\SignalementRepository;
-use App\Repository\SuiviRepository;
 use App\Repository\TerritoryRepository;
 use App\Repository\UserRepository;
 use App\Service\DashboardTabPanel\Kpi\TabCountKpi;
@@ -30,25 +35,35 @@ class TabDataManagerTest extends WebTestCase
 {
     protected MockObject&Security $security;
     protected MockObject&JobEventRepository $jobEventRepository;
-    protected MockObject&SuiviRepository $suiviRepository;
     protected MockObject&TerritoryRepository $territoryRepository;
     protected MockObject&UserRepository $userRepository;
-    protected MockObject&PartnerRepository $partnerRepository;
     protected MockObject&SignalementRepository $signalementRepository;
     protected MockObject&TabCountKpiBuilder $tabCountKpiBuilder;
     protected MockObject&SignalementsSansAffectationAccepteeQuery $signalementsSansAffectationAccepteeQuery;
+    protected MockObject&DossiersQuery $dossiersQuery;
+    protected MockObject&DossiersActiviteRecenteQuery $dossiersActiviteRecenteQuery;
+    protected MockObject&DossiersAvecRelanceSansReponseQuery $dossiersAvecRelanceSansReponseQuery;
+    protected MockObject&DossiersSuivisUsagerQuery $dossiersSuivisUsagerQuery;
+    protected MockObject&DossiersSansSuivisPartenaireQuery $dossiersSansSuivisPartenaireQuery;
+    protected MockObject&DossiersUndeliverableEmailQuery $dossiersUndeliverableEmailQuery;
+    protected MockObject&KpiQuery $kpiQuery;
 
     protected function setUp(): void
     {
         $this->security = $this->createMock(Security::class);
         $this->jobEventRepository = $this->createMock(JobEventRepository::class);
-        $this->suiviRepository = $this->createMock(SuiviRepository::class);
         $this->territoryRepository = $this->createMock(TerritoryRepository::class);
         $this->userRepository = $this->createMock(UserRepository::class);
-        $this->partnerRepository = $this->createMock(PartnerRepository::class);
         $this->signalementRepository = $this->createMock(SignalementRepository::class);
         $this->tabCountKpiBuilder = $this->createMock(TabCountKpiBuilder::class);
         $this->signalementsSansAffectationAccepteeQuery = $this->createMock(SignalementsSansAffectationAccepteeQuery::class);
+        $this->dossiersQuery = $this->createMock(DossiersQuery::class);
+        $this->dossiersActiviteRecenteQuery = $this->createMock(DossiersActiviteRecenteQuery::class);
+        $this->dossiersAvecRelanceSansReponseQuery = $this->createMock(DossiersAvecRelanceSansReponseQuery::class);
+        $this->dossiersSuivisUsagerQuery = $this->createMock(DossiersSuivisUsagerQuery::class);
+        $this->dossiersSansSuivisPartenaireQuery = $this->createMock(DossiersSansSuivisPartenaireQuery::class);
+        $this->dossiersUndeliverableEmailQuery = $this->createMock(DossiersUndeliverableEmailQuery::class);
+        $this->kpiQuery = $this->createMock(KpiQuery::class);
     }
 
     private function getTabDataManager(): TabDataManager
@@ -56,13 +71,17 @@ class TabDataManagerTest extends WebTestCase
         return new TabDataManager(
             $this->security,
             $this->jobEventRepository,
-            $this->suiviRepository,
             $this->territoryRepository,
             $this->userRepository,
-            $this->partnerRepository,
-            $this->signalementRepository,
             $this->tabCountKpiBuilder,
             $this->signalementsSansAffectationAccepteeQuery,
+            $this->dossiersQuery,
+            $this->dossiersActiviteRecenteQuery,
+            $this->dossiersAvecRelanceSansReponseQuery,
+            $this->dossiersSuivisUsagerQuery,
+            $this->dossiersSansSuivisPartenaireQuery,
+            $this->dossiersUndeliverableEmailQuery,
+            $this->kpiQuery
         );
     }
 
@@ -75,7 +94,7 @@ class TabDataManagerTest extends WebTestCase
         $signalementStatus = SignalementStatus::ACTIVE;
         $suiviCategory = SuiviCategory::MESSAGE_PARTNER;
 
-        $this->suiviRepository->method('findLastSignalementsWithUserSuivi')->willReturn([
+        $this->dossiersActiviteRecenteQuery->method('findLastSignalementsWithUserSuivi')->willReturn([
             [
                 'nomOccupant' => 'Dupont',
                 'prenomOccupant' => 'Jean',
@@ -137,7 +156,7 @@ class TabDataManagerTest extends WebTestCase
     {
         $countPartnerDto = $this->createMock(CountPartner::class);
         $countPartnerDto->method('getNonNotifiables')->willReturn(7);
-        $this->partnerRepository->method('countPartnerNonNotifiables')->willReturn($countPartnerDto);
+        $this->kpiQuery->method('countPartnerNonNotifiables')->willReturn($countPartnerDto);
 
         $tabDataManager = $this->getTabDataManager();
 
@@ -147,7 +166,7 @@ class TabDataManagerTest extends WebTestCase
 
     public function testCountPartenairesInterfacesReturnsCount(): void
     {
-        $this->partnerRepository->method('countPartnerInterfaces')->willReturn(5);
+        $this->kpiQuery->method('countPartnerInterfaces')->willReturn(5);
 
         $tabDataManager = $this->getTabDataManager();
 
@@ -188,11 +207,11 @@ class TabDataManagerTest extends WebTestCase
         $expectedCount = 0;
         $params = new TabQueryParameters(null, null);
 
-        $this->signalementRepository
+        $this->dossiersQuery
             ->method('findDossiersNoAgentFrom')
             ->with(AffectationStatus::ACCEPTED, $params)
             ->willReturn($expectedDossiers);
-        $this->signalementRepository
+        $this->dossiersQuery
             ->method('countDossiersNoAgentFrom')
             ->with(AffectationStatus::ACCEPTED, $params)
             ->willReturn($expectedCount);
@@ -209,11 +228,11 @@ class TabDataManagerTest extends WebTestCase
         $expectedDossiers = [['id' => 1], ['id' => 2]];
         $expectedCount = 2;
 
-        $this->signalementRepository
+        $this->dossiersQuery
             ->method('findDossiersDemandesFermetureByUsager')
             ->with(null)
             ->willReturn($expectedDossiers);
-        $this->signalementRepository
+        $this->dossiersQuery
             ->method('countDossiersDemandesFermetureByUsager')
             ->with(null)
             ->willReturn($expectedCount);
@@ -230,7 +249,7 @@ class TabDataManagerTest extends WebTestCase
         /** @var MockObject&User $user */
         $user = $this->createMock(User::class);
         $this->security->method('getUser')->willReturn($user);
-        $this->suiviRepository->method('findSuivisUsagersWithoutAskFeedbackBefore')->with($user)->willReturn([
+        $this->dossiersSuivisUsagerQuery->method('findSuivisUsagersWithoutAskFeedbackBefore')->with($user)->willReturn([
             [
                 'nomOccupant' => 'Martin',
                 'prenomOccupant' => 'Alice',
@@ -243,7 +262,7 @@ class TabDataManagerTest extends WebTestCase
                 'uuid' => 'uuid-456',
             ],
         ]);
-        $this->suiviRepository->method('countSuivisUsagersWithoutAskFeedbackBefore')->with($user)->willReturn(1);
+        $this->dossiersSuivisUsagerQuery->method('countSuivisUsagersWithoutAskFeedbackBefore')->with($user)->willReturn(1);
 
         $tabDataManager = $this->getTabDataManager();
         $result = $tabDataManager->getMessagesUsagersNouveauxMessages();
@@ -262,12 +281,12 @@ class TabDataManagerTest extends WebTestCase
         $expectedCount = 1;
         $params = new TabQueryParameters(null, null);
 
-        $this->signalementRepository
-            ->method('findSignalementsAvecRelancesSansReponse')
+        $this->dossiersAvecRelanceSansReponseQuery
+            ->method('findSignalements')
             ->with($params)
             ->willReturn($expectedDossiers);
-        $this->signalementRepository
-            ->method('countSignalementsAvecRelancesSansReponse')
+        $this->dossiersAvecRelanceSansReponseQuery
+            ->method('countSignalements')
             ->with($params)
             ->willReturn($expectedCount);
         $tabDataManager = $this->getTabDataManager();
@@ -283,7 +302,7 @@ class TabDataManagerTest extends WebTestCase
         /** @var MockObject&User $user */
         $user = $this->createMock(User::class);
         $this->security->method('getUser')->willReturn($user);
-        $this->suiviRepository->method('findSuivisPostCloture')->with($user)->willReturn([
+        $this->dossiersSuivisUsagerQuery->method('findSuivisPostCloture')->with($user)->willReturn([
             [
                 'nomOccupant' => 'Durand',
                 'prenomOccupant' => 'Paul',
@@ -297,7 +316,7 @@ class TabDataManagerTest extends WebTestCase
                 'uuid' => 'uuid-789',
             ],
         ]);
-        $this->suiviRepository->method('countSuivisPostCloture')->with($user)->willReturn(1);
+        $this->dossiersSuivisUsagerQuery->method('countSuivisPostCloture')->with($user)->willReturn(1);
 
         $tabDataManager = $this->getTabDataManager();
         $result = $tabDataManager->getMessagesUsagersMessageApresFermeture();
@@ -315,11 +334,11 @@ class TabDataManagerTest extends WebTestCase
         $expectedCount = 1;
         $params = new TabQueryParameters(null, null);
 
-        $this->signalementRepository
+        $this->dossiersQuery
             ->method('findDossiersFermePartenaireTous')
             ->with($params)
             ->willReturn($expectedDossiers);
-        $this->signalementRepository
+        $this->dossiersQuery
             ->method('countDossiersFermePartenaireTous')
             ->with($params)
             ->willReturn($expectedCount);
@@ -337,11 +356,11 @@ class TabDataManagerTest extends WebTestCase
         $expectedCount = 1;
         $params = new TabQueryParameters(null, null);
 
-        $this->signalementRepository
+        $this->dossiersQuery
             ->method('findDossiersFermePartenaireCommune')
             ->with($params)
             ->willReturn($expectedDossiers);
-        $this->signalementRepository
+        $this->dossiersQuery
             ->method('countDossiersFermePartenaireCommune')
             ->with($params)
             ->willReturn($expectedCount);
@@ -358,7 +377,7 @@ class TabDataManagerTest extends WebTestCase
         /** @var MockObject&User $user */
         $user = $this->createMock(User::class);
         $this->security->method('getUser')->willReturn($user);
-        $this->suiviRepository->method('findSuivisUsagerOrPoursuiteWithAskFeedbackBefore')->with($user)->willReturn([
+        $this->dossiersSuivisUsagerQuery->method('findSuivisUsagerOrPoursuiteWithAskFeedbackBefore')->with($user)->willReturn([
             [
                 'nomOccupant' => 'Lemoine',
                 'prenomOccupant' => 'Claire',
@@ -372,7 +391,7 @@ class TabDataManagerTest extends WebTestCase
                 'uuid' => 'uuid-999',
             ],
         ]);
-        $this->suiviRepository->method('countSuivisUsagerOrPoursuiteWithAskFeedbackBefore')->with($user)->willReturn(1);
+        $this->dossiersSuivisUsagerQuery->method('countSuivisUsagerOrPoursuiteWithAskFeedbackBefore')->with($user)->willReturn(1);
 
         $tabDataManager = $this->getTabDataManager();
 
@@ -392,7 +411,7 @@ class TabDataManagerTest extends WebTestCase
         /** @var MockObject&User $user */
         $user = $this->createMock(User::class);
         $this->security->method('getUser')->willReturn($user);
-        $this->signalementRepository->method('findSignalementsSansSuiviPartenaireDepuis60Jours')->with($user)->willReturn([
+        $this->dossiersSansSuivisPartenaireQuery->method('findSignalements')->with($user)->willReturn([
             [
                 'nomOccupant' => 'Lemoine',
                 'prenomOccupant' => 'Claire',
@@ -408,7 +427,7 @@ class TabDataManagerTest extends WebTestCase
                 'uuid' => 'uuid-999',
             ],
         ]);
-        $this->signalementRepository->method('countSignalementsSansSuiviPartenaireDepuis60Jours')->with($user)->willReturn(1);
+        $this->dossiersSansSuivisPartenaireQuery->method('countSignalements')->with($user)->willReturn(1);
 
         $tabDataManager = $this->getTabDataManager();
 
@@ -479,7 +498,7 @@ class TabDataManagerTest extends WebTestCase
 
         $this->security->method('getUser')->willReturn($user);
         $params = new TabQueryParameters(null, null);
-        $this->suiviRepository->method('findLastSignalementsWithOtherUserSuivi')
+        $this->dossiersActiviteRecenteQuery->method('findLastSignalementsWithOtherUserSuivi')
             ->with($user, $params, 11)
             ->willReturn([
                 [
